+++ lib/plugin/header_menu.dart
@@ -7,6 +7,7 @@ import '../slick_core.dart' as core;
 import 'dart:async';
 import 'package:logging/logging.dart';
 Logger _log = new Logger('log.headermenu');
+typedef void MenuFun(MouseEvent e);
 /***
    * A plugin to add drop-down menus to column headers.
    *
@@ -71,11 +72,11 @@ Logger _log = new Logger('log.headermenu');
    *    buttonImage:      a url to the menu button image (default '../images/down.gif')
    */
 class HeaderMenu  extends IPlugin{
-    
+
     Map _opt;
     //Column _parent;
     HeaderMenu(this._opt){
-      
+
     }
     /**
      * We can add or modify the menu here, or cancel it by returning false
@@ -87,7 +88,7 @@ class HeaderMenu  extends IPlugin{
     String get buttonCssClass => _opt['buttonCssClass'];
     String get buttonImage => _opt['buttonImage'];
     String get tooltip => _opt['tooltip'];
-    
+
     SlickGrid _grid;
     core.EventHandler _handler = new core.EventHandler();
     Element _$menu;
@@ -137,7 +138,7 @@ class HeaderMenu  extends IPlugin{
      * [args] : {"node": $header,
                  "column": Column}
      */
-    handleHeaderCellRendered(e, args) {
+    handleHeaderCellRendered(core.EventData e,Map args) {
       //_parent = args['column'];
 //      Map menu = {};
       Column column=args['column'];
@@ -154,7 +155,7 @@ class HeaderMenu  extends IPlugin{
         if (tooltip!=null) {
           $el.attributes["title"]= tooltip;
         }
-        $el.onClick.listen(_showMenuFun(_showMenu,args['column']));
+        $el.onClick.listen(_showMenuFun(_showMenu,args['column']) );
         (args['node'] as Element).append($el);
     }
 
@@ -167,15 +168,15 @@ class HeaderMenu  extends IPlugin{
       }
     }
 
-    Function _showMenuFun(Function f, Column column)=>(e)=> f(column,e);
-    
+    MenuFun _showMenuFun(Function f, Column column)=>(e)=> f(column,e);
+
     _showMenu(Column column,MouseEvent e) {
      // var menu = $menuButton.data("menu");
      // var columnDef = $menuButton.data("column");
       if(column.header.length==0) return;
       // Let the user modify the menu or cancel altogether,
       // or provide alternative menu implementation.
-      List<MenuItem> menuList = column.header['menu']['items'].map((_) => new MenuItem(_)).toList();
+      List<MenuItem> menuList = column.header['menu']['items'].map((_) => new MenuItem(_)).toList() as List<MenuItem>;
       if (onBeforeMenuShow.notify({
           "grid": _grid,
           "column": column,
@@ -190,7 +191,7 @@ class HeaderMenu  extends IPlugin{
         _grid.container.children.add(_$menu);
       }
       _$menu.children.clear();
-      
+
 
       // Construct the menu items.
       for (var i = 0; i < menuList.length; i++) {
@@ -243,10 +244,10 @@ class HeaderMenu  extends IPlugin{
       e.stopPropagation();
     }
 
-    _handleMenuItemClickFun(Function f,Column column,MenuItem item) => (e) => f(column,item,e);
+  MenuFun  _handleMenuItemClickFun(Function f,Column column,MenuItem item) => (MouseEvent e) => f(column,item,e);
     _menuItemClick(Column column,MenuItem item,MouseEvent e) {
       _log.finest('click:${column.name} ${item.command}');
-      
+
       if (item.disabled) {
         return;
       }
@@ -302,7 +303,7 @@ class MenuItem{
   void set title          (_) => _opt['title']          = _;
   void set command        (_) => _opt['command']        = _;
   void set iconCssClass   (_) => _opt['iconCssClass']   = _;
-  void set iconImage      (_) => _opt['iconImage']      = _;  
+  void set iconImage      (_) => _opt['iconImage']      = _;
   void set tooltip        (_) => _opt['tooltip']        = _;
   set disabled       (bool _) => _opt['disabled']       = _;
 }