+++ web/example/Bs3.dart
@@ -36,7 +36,7 @@ List makeData(int len) {
 
 cj.SlickGrid prepareGrid() {
   Element el = querySelector('#grid');
-  List column = [
+  List<cj.Column> column = [
     new cj.Column.fromMap({'field': "title", 'name': "FIXED", 'sortable': true}),
     new cj.Column.fromMap({'field': "duration", 'name': "A", 'width': 120, 'sortable': true, 'editor': 'IntEditor'}),
     new cj.Column.fromMap({'field': "percent", 'name': 'B', 'sortable': true, 'editor': 'TextEditor'}),