+++ web/example/Formatter.dart
@@ -2,16 +2,17 @@ import 'dart:html';
 import 'dart:convert';
 import 'package:slickdart/slick.dart' as grid;
 import 'dart:math' as math;
+import 'package:slickdart/slick_cell_selection.dart' as cellMode;
 
 void main() {
-  grid.SlickGrid  g=init();
+  grid.SlickGrid g = init();
   g.init();
 //  print (g.$headerScroller.querySelectorAll('.slick-header-column').length);
-  querySelector('#reset').onClick.listen((e){
-    List _data=[];
+  querySelector('#reset').onClick.listen((e) {
+    List _data = [];
     for (var i = 0; i < 50000; i++) {
-      _data.add( {
-        'dtitle':  new math.Random().nextInt(1000).toString(),
+      _data.add({
+        'dtitle': new math.Random().nextInt(1000).toString(),
         'duration': new math.Random().nextInt(1000).toString(),
         'pc': new math.Random().nextInt(100),
         'effortDriven': (i % 5 == 0)
@@ -29,47 +30,66 @@ void main() {
 //  });
 }
 
-grid.SlickGrid init(){
-  Element el =querySelector('#grid');
-  List column = [
-                 new grid.Column.fromMap ({'id': "title", 'name': "Title1", 'field': "dtitle", 'sortable': true,'editor': 'TextEditor' , 'formatter':new SuperFormater()}),
-                 new grid.Column.fromMap ({'width':120,'id': "duration", 'name': "duration", 'field': "duration", 'sortable': true }),
-                 new grid.Column.fromMap ({'id': "%", 'name': "percentComplete", 'field': "pc", 'sortable': true, 'formatter': grid.PercentCompleteBarFormatter }),
-                 new grid.Column.fromMap ({'id': "effort-driven", 'name': "Effort Driven", 'sortable': false, 'width': 80, 'minWidth': 20, 'maxWidth': 80,
-                   'cssClass': "cell-effort-driven", 'field': "effortDriven", 'formatter': grid.CheckmarkFormatter})
+grid.SlickGrid init() {
+  Element el = querySelector('#grid');
+  List<grid.Column> column = [
+    new grid.Column.fromMap({
+      'id': "title",
+      'name': "Title1",
+      'field': "dtitle",
+      'sortable': true,
+      'editor': 'TextEditor',
+      'formatter': new SuperFormater()
+    }),
+    new grid.Column.fromMap(
+        {'width': 120, 'id': "duration", 'name': "duration", 'field': "duration", 'sortable': true}),
+    new grid.Column.fromMap({
+      'id': "%",
+      'name': "percentComplete",
+      'field': "pc",
+      'sortable': true,
+      'formatter': grid.PercentCompleteBarFormatter
+    }),
+    new grid.Column.fromMap({
+      'id': "effort-driven",
+      'name': "Effort Driven",
+      'sortable': false,
+      'width': 80,
+      'minWidth': 20,
+      'maxWidth': 80,
+      'cssClass': "cell-effort-driven",
+      'field': "effortDriven",
+      'formatter': grid.CheckmarkFormatter
+    })
 
-                // new grid.Column.fromMap ({'id': "start", 'name': "finish", 'field': "finish"})
-                 ];
-  List data=[];
+    // new grid.Column.fromMap ({'id': "start", 'name': "finish", 'field': "finish"})
+  ];
+  List data = [];
   for (var i = 0; i < 50000; i++) {
-    data.add( {
-      'dtitle':  i.toString(),
+    data.add({
+      'dtitle': i.toString(),
       'duration': new math.Random().nextInt(100).toString(),
       'pc': new math.Random().nextInt(100),
       'effortDriven': (i % 5 == 0)
     });
   }
-  Map opt = {'explicitInitialization': false,
-             'multiColumnSort': true,
-             'editable': true,
-  };
-  grid.SlickGrid sg= new grid.SlickGrid(el,data,column,opt);
+  Map opt = {'explicitInitialization': false, 'multiColumnSort': true, 'editable': true,};
+  grid.SlickGrid sg = new grid.SlickGrid(el, data, column, opt);
 
-  sg.setSelectionModel(new grid.CellSelectionModel(sg.options));
+  sg.setSelectionModel(new cellMode.CellSelectionModel(sg.options));
 
-
-  sg.onSort.subscribe( (e, args) {
+  sg.onSort.subscribe((e, args) {
     var cols = args['sortCols'];
 //{sortCol: {name: Title1, resizable: true, sortable: true, minWidth: 30, rerenderOnResize: false, headerCssClass: null, defaultSortAsc: true, focusable: true, selectable: true, cannotTriggerInsert: false, width: 80, id: title, field: title}, sortAsc: true}
-    data.sort( (dataRow1, dataRow2) {
+    data.sort((dataRow1, dataRow2) {
       for (var i = 0, l = cols.length; i < l; i++) {
         var field = cols[i]['sortCol']['field'];
         var sign = cols[i]['sortAsc'] ? 1 : -1;
         dynamic value1 = dataRow1[field], value2 = dataRow2[field];
-        if(field=='dtitle') {
-          return value1 == value2 ? 0 : (int.parse(value1) > int.parse(value2) ? 1: -1)* sign;
+        if (field == 'dtitle') {
+          return value1 == value2 ? 0 : (int.parse(value1) > int.parse(value2) ? 1 : -1) * sign;
         }
-        var result = (value1 == value2 ? 0 : (value1.compareTo(value2)>0 ? 1 : -1)) * sign;
+        var result = (value1 == value2 ? 0 : (value1.compareTo(value2) > 0 ? 1 : -1)) * sign;
         if (result != 0) {
           return result;
         }
@@ -82,20 +102,22 @@ grid.SlickGrid init(){
   return sg;
 }
 
-class SuperFormater{
-  call(int row,int  cell, dynamic value,grid.Column columnDef,dataContext){
+class SuperFormater {
+  call(int row, int cell, dynamic value, grid.Column columnDef, dataContext) {
     /**demo code for ser/deser */
     var colStr = JSON.encode(columnDef);
     new grid.Column.fromJSON(colStr);
     /** end */
-      return value;
+    return value;
   }
-  toString(){
+
+  toString() {
     return 'SuperFormater';
   }
 }
-SuperFormatter2(int row,int  cell, dynamic value,grid.Column columnDef,dataContext) {
+
+SuperFormatter2(int row, int cell, dynamic value, grid.Column columnDef, dataContext) {
   var colStr = JSON.encode(columnDef.toString());
- // grid.Column col = new grid.Column.fromJSON(colStr);
+  // grid.Column col = new grid.Column.fromJSON(colStr);
   return value;
 }